Centralize workout time helper functions
