+++ Uplift/ViewModels/WorkoutHistoryViewModel.swift
@@ -37,7 +37,7 @@ extension WorkoutHistoryView {
 
             let sortedWorkouts = workouts
                 .compactMap { workout -> (Workout, Date)? in
-                    guard let date = workoutIsoToDate(workout.workoutTime) else { return nil }
+                    guard let date = WorkoutTimeFormatter.isoToDate(workout.workoutTime) else { return nil }
                     return (workout, date)  // allow sorting by date
                 }
                 .sorted { $0.1 > $1.1 } // sort by newest first
@@ -170,52 +170,6 @@ extension WorkoutHistoryView {
         /// Moves the selected month back by one month.
         func prevMonth() {
             selectedMonth = calendar.date(byAdding: .month, value: -1, to: firstOfCurrMonth)!
-        }
-
-        /// Returns the `Date` object from the given workout time in ISO 18601 with timezone format from backend.
-        private func workoutIsoToDate(_ workoutTime: String) -> Date? {
-            let iso = ISO8601DateFormatter()
-            iso.formatOptions = [.withInternetDateTime, .withTimeZone]
-            return iso.date(from: workoutTime)
-        }
-
-        /// Formats the workout time from backend (ISO 8601 with timezone) as `MMM d • h:mm a` in local time.
-        /// Returns the original format if parsing fails.
-        func stringToWorkoutTime(_ workoutTime: String, in tab: WorkoutHistoryTab) -> String {
-            guard let date = workoutIsoToDate(workoutTime) else {
-                Logger.data.critical("Error in WorkoutHistoryViewModel: Formatter unable to parse workout time")
-                return workoutTime
-            }
-
-            let dateFormatter = DateFormatter()
-            dateFormatter.locale = Locale(identifier: "en_US_POSIX")
-            dateFormatter.timeZone = calendar.timeZone
-            dateFormatter.dateFormat = "MMM d"  // e.g. Mar 3
-
-            let timeFormatter = DateFormatter()
-            timeFormatter.locale = Locale(identifier: "en_US_POSIX")
-            timeFormatter.timeZone = calendar.timeZone
-            timeFormatter.dateFormat = "h:mm a" // e.g. 3:10 PM
-
-            switch tab {
-            case .calendar:
-                return "\(timeFormatter.string(from: date)) • \(dateFormatter.string(from: date))"
-            case .list:
-                return "\(dateFormatter.string(from: date)) • \(timeFormatter.string(from: date))"
-            }
-        }
-
-        /// Formats the workout time from backend to a localized relative label (e.g. "today", "yesterday")..
-        /// Returns empty string if parsing fails.
-        func relativeWorkoutTime(_ workoutTime: String) -> String {
-            guard let date = workoutIsoToDate(workoutTime) else {
-                Logger.data.critical("Error in WorkoutHistoryViewModel: Formatter unable to parse workout time")
-                return ""
-            }
-
-            let relativeString = date.formatted(.relative(presentation: .named))
-            guard let first = relativeString.first else { return "" }
-            return String(first.uppercased()) + relativeString.dropFirst()
         }
 
         // TODO: Remove later
